Move read related methods to separate file
